@@ -3,8 +3,9 @@
     "description": "webmachine for PHP",
     "license": "BSD-3-Clause",
     "require": {
-        "php": ">=5.4",
-        "symfony/http-foundation": "~2.6"
+        "php": ">=7.4.*",
+        "symfony/http-foundation": "~2.6",
+        "ext-json": "*"
     },
     "autoload": {
         "psr-4": {
@@ -13,7 +14,7 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.7",
-        "silex/silex": "^1.3"
+        "silex/silex": "^1.3",
+        "phpunit/phpunit": "^9.2"
     }
 }
